requirements.txt
APScheduler==3.9.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
Brotli==1.0.9
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
coverage==6.3.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-celery-beat==2.2.1
django-cloudinary-storage==0.3.0
django-environ==0.8.1
django-heroku==0.3.1
django-redis==5.2.0
django-timezone-field==4.2.3
flake8==4.0.1
html5lib==1.1
idna==3.3
kombu==5.2.4
mccabe==0.6.1
Naked==0.1.31
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydocstyle==6.1.1
pydyf==0.1.2
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.1.4
reportlab==3.6.8
requests==2.27.1
shellescape==3.8.1
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.0.0
wkhtmltopdf==0.2
wrapt==1.14.0
zopfli==0.2.1
